+++ Suspension/src/Suspension.Core/Operations/CalculateReactions/CalculateReactions.cs
@@ -96,8 +96,7 @@ namespace Suspension.Core.Operations.CalculateReactions
                 return response;
             }
 
-            Vector3D appliedForce = Vector3D.Create(request.AppliedForce);
-            if (appliedForce.X == 0 && appliedForce.Y == 0 && appliedForce.Z == 0 || request.AppliedForce == "0,0,0")
+            if (Vector3D.Create(request.AppliedForce).IsZero())
             {
                 response.SetBadRequestError("The applied force must have at least one coordinate different than zero.");
             }