+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Fatigue constitutive equations and operation.
  - Methods 'Abs', 'Sum' and 'Subtract' to class Force.
  - Properties BucklingSafetyFactor and StressSafetyFactor to RunStaticAnalysisResponseData.
+ - Extension method 'IsZero' for Vector3D.
 ### Changed
  - Renamed RunAnalysis operation to RunStaticAnalysis and refactored it.
  - Point3D to receive the point as string at milimeters instead of meters.